riscv: mm: Use a fixed layout for the MM context ID
mainline inclusion
from mainline-6.10
commit f58e5dc45fa93a2f2a0028381b2e06fe74ae9e2c ("riscv: mm: Use a fixed layout for the MM context ID")
category: feature
Link: https://github.com/RVCK-Project/rvck/issues/332

--------------------------------

Currently, the size of the ASID field in the MM context ID dynamically
depends on the number of hardware-supported ASID bits. This requires
reading a global variable to extract either field from the context ID.
Instead, allocate the maximum possible number of bits to the ASID field,
so the layout of the context ID is known at compile-time.

Reviewed-by: Alexandre Ghiti <alexghiti@rivosinc.com>
Signed-off-by: Samuel Holland <samuel.holland@sifive.com>
Link: https://lore.kernel.org/r/20240327045035.368512-11-samuel.holland@sifive.com
Signed-off-by: Palmer Dabbelt <palmer@rivosinc.com>
Signed-off-by: ZhenXing Zhu <zhenxing.zhu@linux.alibaba.com>